Unused entities were removed. PlatformDependentIODispatcher was moved to :Core module.
