+++ Core/build.gradle.kts
@@ -8,7 +8,7 @@ plugins {
 }
 
 group = "io.growthbook.sdk"
-version = "1.0.0"
+version = "1.0.1"
 
 kotlin {
     android {
