+++ Core/src/commonMain/kotlin/com/sdk/growthbook/Dispatcher.kt
@@ -5,4 +5,4 @@ import kotlinx.coroutines.CoroutineDispatcher
 /**
  * Expect IO Dispatcher - from respective iOS & Android counter parts
  */
-internal expect val PlatformDependentIODispatcher: CoroutineDispatcher
+expect val PlatformDependentIODispatcher: CoroutineDispatcher
