+++ Core/src/jvmMain/kotlin/com/sdk/growthbook/Dispatcher.kt
@@ -4,5 +4,5 @@ import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 
 // 1
-internal actual val PlatformDependentIODispatcher: CoroutineDispatcher =
+actual val PlatformDependentIODispatcher: CoroutineDispatcher =
     Dispatchers.IO
