+++ OkHttpNetworkDispatcher/build.gradle.kts
@@ -21,38 +21,28 @@ kotlin {
         }
     }
 
-    val ktorVersion = "2.1.2"
     sourceSets {
         val commonMain by getting {
             dependencies {
                 implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.2.0")
-
-                api("io.ktor:ktor-client-core:$ktorVersion")
-                implementation("io.ktor:ktor-client-content-negotiation:$ktorVersion")
-                implementation("io.ktor:ktor-serialization-kotlinx-json:$ktorVersion")
-                implementation("com.squareup.okhttp3:okhttp:4.11.0")
-
-                // Server Sent Events
-                implementation("com.squareup.okhttp3:okhttp-sse:4.11.0")
-                implementation("com.launchdarkly:okhttp-eventsource:2.5.0")
+                api("com.squareup.okhttp3:okhttp:4.9.0")
+                implementation("com.squareup.okhttp3:okhttp-sse:4.9.0")
 
                 // https://mvnrepository.com/artifact/com.google.code.gson/gson
-                implementation("com.google.code.gson:gson:2.11.0")
+                //implementation("com.google.code.gson:gson:2.11.0")
 
                 implementation("io.growthbook.sdk:Core:1.0.0")
             }
         }
         val androidMain by getting {
             dependencies {
-                implementation("io.ktor:ktor-client-android:$ktorVersion")
+                implementation("com.squareup.okhttp3:okhttp:4.9.0")
             }
         }
         val androidTest by getting {
             dependencies {
                 implementation(kotlin("test-junit"))
-                implementation("io.ktor:ktor-client-mock:$ktorVersion")
-                implementation("com.squareup.okhttp3:mockwebserver:4.12.0")
-
+                implementation("com.squareup.okhttp3:mockwebserver:4.9.0")
             }
         }
     }
