--- OkHttpNetworkDispatcher/src/androidTest/kotlin/com/sdk/growthbook/OkHttpGBNetworkDispatcherTest.kt
@@ -1,18 +1,0 @@
-package com.sdk.growthbook
-
-
-import org.junit.Test
-
-import org.junit.Assert.assertTrue
-
-private const val FEATURES_ENDPOINT = "/api/features/"
-
-class OkHttpGBNetworkDispatcherTest {
-
-    @Test
-    fun `test successful get request`() {
-
-
-        assertTrue(true)
-    }
-}
