--- OkHttpNetworkDispatcher/src/commonMain/kotlin/com/sdk/growthbook/utils/another/SSEHandler.kt
@@ -1,10 +1,0 @@
-package com.sdk.growthbook.utils.another
-
-import com.launchdarkly.eventsource.MessageEvent
-
-interface SSEHandler {
-    fun onSSEConnectionOpened()
-    fun onSSEConnectionClosed()
-    fun onSSEEventReceived(event: String, messageEvent: MessageEvent)
-    fun onSSEError(t: Throwable)
-}
